% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transfer_adapter.R
\name{basic_upload}
\alias{basic_upload}
\title{Do a file upload using basic transfer adapter}
\usage{
basic_upload(file_path, upload_specs)
}
\arguments{
\item{file_path}{character. The path of the file to be uploaded.}

\item{upload_specs}{A list of upload specs returned by the batch call.}
}
\description{
Gets the upload action from the upload specs and makes a PUT request to
upload the file to the server. After uploading the file it executes
a commit and verify action.
}
