% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{Send a batch request to the LFS server}
\usage{
batch(prefix, objects, token, transfers, headers = c())
}
\arguments{
\item{prefix}{character. The lfs prefix in the form organization/repo}

\item{objects}{list. A list of objects containing oid and size for each one.}

\item{token}{character. An authorization token.}

\item{transfers}{character vector. A vector containing the transfer methods
to be negotiated with the server.}

\item{headers}{named character vector. Values to be added to the headers of
the batch request to the LFS server..}
}
\value{
List with upload specifications for the given objects
}
\description{
Send a batch request to the LFS server
}
