% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_want_digest}
\alias{calculate_want_digest}
\title{Calculates a digest for the uploading part}
\usage{
calculate_want_digest(data, want_digest)
}
\arguments{
\item{data}{raw object.}

\item{want_diggest}{character. A string specifying the digest algorithm to
use. (Currently only contentMD5 is supported)}
}
\value{
caracter. The digest string: for contentMD5 it returns a base64
encoded digest.
}
\description{
Calculates a digest for the uploading part
}
