% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lfs_server_url}
\alias{get_lfs_server_url}
\title{Get the Giftless API url from the LFS_SERVER_URL environment variable}
\usage{
get_lfs_server_url()
}
\value{
Full URL for Giftless API requests
}
\description{
Get the Giftless API url from the LFS_SERVER_URL environment variable
}
