% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{lfs_upload}
\alias{lfs_upload}
\title{Upload a file to LFS storage}
\usage{
lfs_upload(
  file_path,
  repo,
  dataset,
  token,
  transfers = c("multipart-basic", "basic"),
  headers = c()
)
}
\arguments{
\item{file_path}{character. The path of the file to be uploaded.}

\item{repo}{character. The name of repository where the dataset lives.}

\item{dataset}{character. The name of the dataset where the file is gonna
be uploaded.}

\item{token}{character. Authorization token.}

\item{transfers}{character vector. A vector containing the transfer methods
to be negotiated with the server.}

\item{headers}{named character vector. Values to be added to the headers of
the batch request to the LFS server.}
}
\value{
list with sha256 and size of the object
}
\description{
Upload a file to LFS storage
}
