% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipart_transfer_adapter.R
\name{multipart_upload}
\alias{multipart_upload}
\title{Do a file upload using basic transfer adapter}
\usage{
multipart_upload(file_path, upload_specs)
}
\arguments{
\item{file_path}{character. The path of the file to be uploaded.}

\item{upload_specs}{A list of upload specs returned by the batch call.}
}
\description{
Gets the upload action from the upload specs and makes a PUT request to the
server for each part specified. After uploading all the parts it executes
a commit and verify action.
}
