% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_response}
\alias{parse_response}
\title{Parse the content of the LFS server response}
\usage{
parse_response(resp)
}
\arguments{
\item{resp}{response object.}
}
\value{
A parsed response.
}
\description{
Parse the content of the LFS server response
}
