% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upload_part}
\alias{upload_part}
\title{Uploads a file part}
\usage{
upload_part(file_path, href, pos, size, want_digest = NULL)
}
\arguments{
\item{file_path}{character. string naming a file}

\item{href}{character. A signed URL to upload the file to}

\item{pos}{numeric. A file position (relative to the origin specified by
origin), or NA.}

\item{size}{numeric. The (maximal) number of records to be read.}

\item{want_diggest}{character. A string specifying the digest algorithm to
use. (Currently only contentMD5 is supported)}
}
\description{
Uploads a file part to a signed_url given size and initial position. This
method will read a chunk of the file using readBin. If want_digest is set it
will also calculate the digest and send it as a part of the header (only
MD5 is supported)
}
